+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.5.22
 - Added oredict, blocking, processing, ore inputs access to OpenComputers API (raoulvdberge)
+- Added shortcut to clear Grid crafting matrix (CTRL+X) (raoulvdberge)
 - The Crafter can now only store 1 stack size pattern per slot (raoulvdberge)
 - You can now re-insert a Pattern in the pattern output slot in the Pattern Grid to modify an existing pattern (raoulvdberge)
 - Fixed not being able to use JEI R and U keys on Grid with tabs (raoulvdberge)